Version 4.12

+++ pcileech/umd.c
@@ -25,7 +25,7 @@ VOID Action_UmdPsList()
     PDWORD pdwPIDs = NULL;
     PVMMDLL_PROCESS_INFORMATION pProcInfo = NULL;
     // 1: Initialize MemProcFS/vmm.dll
-    if(!(pdwPIDs = LocalAlloc(LMEM_ZEROINIT, cPIDs * sizeof(DWORD)))) { goto fail; }
+    if(!(pdwPIDs = LocalAlloc(LMEM_ZEROINIT, (SIZE_T)(cPIDs * sizeof(DWORD))))) { goto fail; }
     if(!(pProcInfo = LocalAlloc(0, sizeof(VMMDLL_PROCESS_INFORMATION)))) { goto fail; }
     if(!Vmmx_Initialize(FALSE, FALSE)) {
         printf("UMD: Failed initializing required MemProcFS/vmm.dll\n");
@@ -35,7 +35,7 @@ VOID Action_UmdPsList()
     if(!VMMDLL_PidList(pdwPIDs, &cPIDs)) {
         printf("UMD: Failed list PIDs.\n");
     } else {
-        qsort(pdwPIDs, cPIDs, sizeof(DWORD), UmdCompare32);
+        qsort(pdwPIDs, (SIZE_T)cPIDs, sizeof(DWORD), UmdCompare32);
         for(i = 0; i < cPIDs; i++) {
             ZeroMemory(pProcInfo, sizeof(VMMDLL_PROCESS_INFORMATION));
             pProcInfo->magic = VMMDLL_PROCESS_INFORMATION_MAGIC;
